requirements.txt
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
groq==0.9.0
requests==2.31.0
httpx==0.27.0
aiohttp==3.11.11
pytz==2024.1
sqlitedict==2.1.0
pillow>=10.3.0
pymupdf==1.24.1
faiss-cpu==1.8.0
google-generativeai==0.8.3
numpy==1.26.4
pilmoji==2.0.5
emoji==2.15.0
